recipes-firmware: Smartcam
bit-file :using a local ip with gasket for vcu,
dtsi-file :the string 'skip isolation' let the driver know isolation
exists in hw. this is a temporty fix and not final solution.
xclbin-file : new xclbin to meet timing closure by swapping PS ports in VCU
This change in smartcam is to keep consistency with aa2 local ip